+++ pyproject.toml
@@ -1,0 +1,22 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "repler"
+version = "1.0.0"
+authors = [
+  { name="Alexander Barzilai", email="Alex74zel@gmail.com" },
+]
+description = "The tool keeps you project envieronment clean by delete or archive files with specified extensions."
+readme = "README.md"
+requires-python = ">=3.7"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+
+[project.urls]
+"Homepage" = "https://github.com/abarzilai/report-handler/tree/main"
+"Bug Tracker" = "https://github.com/abarzilai/report-handler/issues"
